improve: 公開中の映画のウィジェットを独立させる
